+++ tools/luban-provisioner/templates/source/luban-dagster-dbt-starrocks-code-location-source-template/{{cookiecutter.app_name}}/src/{{cookiecutter.package_name}}/schedules/dbt_config.py
@@ -2,19 +2,27 @@ from .lib.dbt_schedule_presets import daily_at, hourly_at
 
 
 DBT_SCHEDULE_SPECS = [
-    daily_at(name="daily_facts_schedule", job_name="dbt_daily_customer_facts_job", hour=1, minute=0, enabled=True),
+    daily_at(
+        name="daily_facts_schedule",
+        job_name="dbt_daily_customer_facts_job",
+        lookback_days=0,
+        hour=1,
+        minute=0,
+        enabled=True,
+    ),
     daily_at(
         name="finalize_orders_daily_schedule",
-        job_name="dbt_finalize_orders_daily_job",
+        job_name="dbt_orders_daily_job",
+        lookback_days=1,
         hour=1,
         minute=5,
         enabled=True,
     ),
     hourly_at(
-        name="intraday_orders_daily_schedule",
-        job_name="dbt_intraday_orders_daily_job",
-        minute=0,
+        name="orders_hourly_schedule",
+        job_name="dbt_orders_hourly_job",
+        lookback_hours=2,
+        minute=10,
         enabled=False,
     ),
 ]
-