+++ tools/luban-provisioner/templates/source/luban-dagster-dbt-starrocks-code-location-source-template/{{cookiecutter.app_name}}/src/{{cookiecutter.package_name}}/schedules/observations.py
@@ -1,3 +1,5 @@
+import os
+
 from dagster import ScheduleDefinition
 
 from ..jobs.observations import observe_sources_job
@@ -5,7 +7,6 @@ from ..jobs.observations import observe_sources_job
 
 observe_sources_schedule = ScheduleDefinition(
     name="observe_sources_schedule",
-    cron_schedule="* * * * *",
+    cron_schedule=os.getenv("LUBAN_OBSERVE_SOURCES_CRON", "*/5 * * * *"),
     job=observe_sources_job,
 )
-