+++ tools/luban-provisioner/templates/source/luban-dagster-dbt-starrocks-code-location-source-template/{{cookiecutter.app_name}}/src/{{cookiecutter.package_name}}/sensors/__init__.py
@@ -9,4 +9,3 @@ automation_condition_sensor = dg.AutomationConditionSensorDefinition(
 
 
 sensors = [automation_condition_sensor]
-